api/search-templates/requirements.txt
openai==1.14.0
requests==2.32.3
beautifulsoup4==4.13.3
python-dotenv==1.0.1 